profileFace/requirements.txt
absl-py==0.8.0
astor==0.8.0
awscli==1.16.140
botocore==1.12.130
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.4.0
docutils==0.14
gast==0.3.2
google-pasta==0.1.7
graphviz==0.13
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
jmespath==0.9.4
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
moviepy==1.0.0
mtcnn==0.0.9
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.2.2
proglog==0.1.9
protobuf==3.9.1
pycocotools==2.0
pydot==1.4.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.21.3
scipy==1.1.0
six==1.12.0
tensorboard==1.11.0
tensorflow-gpu @ file:///home/zhex/package/tensorflow_gpu-1.11.0-cp36-cp36m-manylinux1_x86_64.whl
termcolor==1.1.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.36.1
urllib3 @ file:///home/zhex/package/urllib3-1.24.3-py2.py3-none-any.whl
Werkzeug==0.16.0
wrapt==1.11.2
